fix: move factory fixture smoke test under backend-size limit
Relocate TestMinimalFactoryConfig_WriteReadSmoke to factoryfixtures so
factory_test_helpers_test.go stays under the 1000-line CI cap.
