+++ docs/internal/processes/factory-workstation-relevant-files.md
@@ -78,7 +78,7 @@ work inputs.
 - Projection and circuit-breaker fixtures that inject runtime workstation maps must key those maps by the authored transition `Name` string, even when projected constraint IDs or token failure history still reference `Transition.ID`; these seams intentionally use different identifiers for lookup versus reporting.
 - When a peer-input guard needs boundary-to-runtime regression coverage, prefer one checked-in fixture under `tests/functional_test/testdata/` that carries `factory.json` plus worker and workstation `AGENTS.md`, and reuse that fixture for generated-boundary, runtime-load, and functional dispatch assertions instead of rebuilding the config inline in each test.
 - When a functional runtime test scaffolds a split-config factory with `tests/functional/internal/support.ScaffoldFactory(...)`, add worker `workers/<name>/AGENTS.md` files explicitly for any dispatched model workers; the helper seeds workstation `AGENTS.md` files but does not author worker configs or stop tokens.
-- Canonical minimal `factory.json` bootstrap for tests lives in `pkg/testutil/factoryfixtures` as `MinimalFactoryConfig()` and `WriteFactoryJSON()`; `pkg/service/factory_test_helpers_test.go` exposes thin `minimalFactoryConfig()` / `writeFactoryJSON()` wrappers for in-package tests and `TestMinimalFactoryConfig_WriteReadSmoke`. Service subpackages (`opencodeagenttests`, `replaytests`, `runtimelogtests`, `modeltests`) and `pkg/replay/configtests` import `factoryfixtures` directly—do not duplicate map or write logic.
+- Canonical minimal `factory.json` bootstrap for tests lives in `pkg/testutil/factoryfixtures` as `MinimalFactoryConfig()`, `WriteFactoryJSON()`, and `TestMinimalFactoryConfig_WriteReadSmoke` in `minimal_test.go`; `pkg/service/factory_test_helpers_test.go` exposes thin `minimalFactoryConfig()` / `writeFactoryJSON()` wrappers for in-package tests only. Service subpackages (`opencodeagenttests`, `replaytests`, `runtimelogtests`, `modeltests`) and `pkg/replay/configtests` import `factoryfixtures` directly—do not duplicate map or write logic.
 - The generated OpenAPI factory/load boundary now accepts workstation `onContinue`, `onRejection`, and `onFailure` only as arrays; do not reintroduce single-object wrapping in `pkg/config/openapi_factory.go`, and treat any checked-in single-object authoring as fixture drift to canonicalize.
 - When public workstation topology is converted between `WorkstationIO` arrays and internal place-id slices, preserve authored slice order end to end; sorting those routes during replay or projection silently changes first-match routing semantics even when the route count survives.
 - When workstation `AGENTS.md` route families such as `onContinue`, `onRejection`, or `onFailure` are arrays, validate them with the same list-walking alias helpers used for `inputs` and `outputs`; otherwise retired-field errors lose their item index and nested route entries drift past authored-boundary checks.
